git-svn-id: https://svn.code.sf.net/p/pinmame/code/branches/unix@3712 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/windows/misc.c
@@ -94,7 +94,7 @@ void fprint_colums(FILE *f, const char *text1, const char *text2)
    /* initialize our arrays */
    text[0] = text1;
    text[1] = text2;
-   width[0] = cols * 0.4;
+   width[0] = (int)((double)cols * 0.4);
    width[1] = cols - width[0];
 
    while(!done)
@@ -118,7 +118,7 @@ void fprint_colums(FILE *f, const char *text1, const char *text2)
             }
 
          /* if we don't have enough space, break at the first ' ' or '\n' */
-         if(to_print < strlen(text[i]))
+         if((size_t)to_print < strlen(text[i]))
          {
            while(to_print && (text[i][to_print] != ' ') &&
               (text[i][to_print] != '\n'))
